.roo/templates/ai-ml-project/requirements.txt
# Production dependencies
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0
joblib>=1.0.0,<2.0.0
pyyaml>=5.4.0,<6.0.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=3.5.0,<4.0.0
mlflow>=1.20.0,<2.0.0

# Development dependencies
pytest>=6.2.0,<7.0.0
pytest-cov>=2.12.0,<3.0.0
pytest-html>=3.1.0,<4.0.0
pytest-xdist>=2.3.0,<3.0.0
pytest-mock>=3.6.0,<4.0.0
black>=21.7b0,<22.0.0
flake8>=3.9.0,<4.0.0
mypy>=0.910,<1.0.0
bandit>=1.7.0,<2.0.0
sphinx>=4.1.0,<5.0.0
sphinx-rtd-theme>=0.5.0,<1.0.0
jupyter>=1.0.0,<2.0.0
locust>=2.0.0,<3.0.0

# Documentation dependencies
docutils>=0.17.0,<1.0.0
sphinx-autodoc-typehints>=1.12.0,<2.0.0
sphinx-copybutton>=0.4.0,<1.0.0

# Testing dependencies
factory-boy>=3.2.0,<4.0.0
freezegun>=1.1.0,<2.0.0
responses>=0.13.0,<1.0.0

# Development tools
pre-commit>=2.13.0,<3.0.0
watchdog>=2.1.0,<3.0.0

# Optional dependencies
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Security
cryptography>=3.4.0,<4.0.0

# Performance
aiofiles>=0.7.0,<1.0.0

# Utilities
python-dotenv>=0.18.0,<1.0.0
click>=8.0.0,<9.0.0
rich>=10.7.0,<11.0.0

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0

# Caching
cachetools>=4.2.0,<5.0.0

# Logging
structlog>=21.1.0,<22.0.0

# API documentation
httpx>=0.18.0,<1.0.0

# Data validation
cerberus>=1.3.0,<2.0.0